Use inline tests with `pytest`, relative imports. All tests pass.
